Refactor `test_text`
